+++ documentdb_tests/compatibility/tests/system/diagnostic/commands/lockInfo/test_lockInfo_argument_validation.py
@@ -1,6 +1,7 @@
 """Tests for lockInfo command argument validation.
 
-Verifies that lockInfo accepts any BSON type for the command field value.
+Verifies that lockInfo accepts any BSON type for the command field value and
+that successive calls succeed (point-in-time snapshot behaviour).
 """
 
 import pytest
@@ -17,6 +18,7 @@ from documentdb_tests.framework.bson_type_validator import (
 from documentdb_tests.framework.executor import execute_admin_command
 from documentdb_tests.framework.parametrize import pytest_params
 from documentdb_tests.framework.property_checks import Eq
+from documentdb_tests.framework.test_constants import FLOAT_INFINITY
 
 pytestmark = pytest.mark.admin
 
@@ -40,7 +42,7 @@ def test_lockInfo_accepts_any_type(collection, bson_type, sample_value, spec):
     assertProperties(result, {"ok": Eq(1.0)}, msg=spec.msg, raw_res=True)
 
 
-INT_EDGE_TESTS: list[DiagnosticTestCase] = [
+VALUE_EDGE_TESTS: list[DiagnosticTestCase] = [
     DiagnosticTestCase(
         id="int_zero",
         command={"lockInfo": 0},
@@ -53,11 +55,24 @@ INT_EDGE_TESTS: list[DiagnosticTestCase] = [
         checks={"ok": Eq(1.0)},
         msg="lockInfo should accept int -1",
     ),
+    DiagnosticTestCase(
+        id="float_infinity",
+        command={"lockInfo": FLOAT_INFINITY},
+        checks={"ok": Eq(1.0)},
+        msg="lockInfo should accept float infinity as the command field value",
+    ),
 ]
 
 
-@pytest.mark.parametrize("test", pytest_params(INT_EDGE_TESTS))
-def test_lockInfo_accepts_int_edge_values(collection, test):
-    """Verify lockInfo accepts zero and negative int values for the command field."""
+@pytest.mark.parametrize("test", pytest_params(VALUE_EDGE_TESTS))
+def test_lockInfo_accepts_value_edge_cases(collection, test):
+    """Verify lockInfo accepts int and float edge values for the command field."""
     result = execute_admin_command(collection, test.command)
     assertProperties(result, test.checks, msg=test.msg, raw_res=True)
+
+
+def test_lockInfo_point_in_time_snapshot(collection):
+    """Test lockInfo result is a point-in-time snapshot (successive calls succeed)."""
+    execute_admin_command(collection, {"lockInfo": 1})
+    result = execute_admin_command(collection, {"lockInfo": 1})
+    assertProperties(result, {"ok": Eq(1.0)}, msg="Successive calls should succeed", raw_res=True)
